coderabbit fixes and changes to fetcher and schedulers
